Renamed and moved tunnel script files at proper locations
